+++ modules/AWSPowerShell/Cmdlets/ConnectWisdomService/AWS.Tools.ConnectWisdomService.Completers.psm1
@@ -101,6 +101,20 @@ $WSDM_Completers = {
             break
         }
 
+        # Amazon.ConnectWisdomService.Relevance
+        "Write-WSDMFeedback/Feedback_Relevance"
+        {
+            $v = "HELPFUL","NOT_HELPFUL"
+            break
+        }
+
+        # Amazon.ConnectWisdomService.TargetType
+        "Write-WSDMFeedback/TargetType"
+        {
+            $v = "RECOMMENDATION","RESULT"
+            break
+        }
+
 
     }
 
@@ -111,7 +125,9 @@ $WSDM_Completers = {
 
 $WSDM_map = @{
     "AssociationType"=@("New-WSDMAssistantAssociation")
+    "Feedback_Relevance"=@("Write-WSDMFeedback")
     "KnowledgeBaseType"=@("New-WSDMKnowledgeBase")
+    "TargetType"=@("Write-WSDMFeedback")
     "Type"=@("New-WSDMAssistant")
 }
 
@@ -187,6 +203,7 @@ $WSDM_SelectMap = @{
                "Get-WSDMKnowledgeBasisList",
                "Get-WSDMResourceTag",
                "Remove-WSDMRecommendationsReceived",
+               "Write-WSDMFeedback",
                "Search-WSDMAssistant",
                "Remove-WSDMKnowledgeBaseTemplateUri",
                "Search-WSDMContent",